==> 03_bgd_hh_drought_join.Rmd <==
---
title: "Match household and drought data"
output: html_document
date: "2024-11-12"
---

```{r libraries}
library("haven")
library("sf")                       # Spatial features
library("raster")                   # Raster data (but also shapefiles)
library("dplyr")
library("exactextractr")            # Zonal statistics using geo-spatial data
```

SPEI data is on a regular 0.5 degress grid (around 50km x 50km at the equator). This means some grid cells are likely to cross the boundaries of unions (administrative level 4). I will use the exactextract package to calculate the mean value of SPEI for each union, based on a raster of SPEI values and a shapefile with boundaries of administrative level 4.

## Match SPEI with BIHS ##

```{r}
# Set working directory
setwd("C:/Users/idabr/OneDrive - Oxford Policy Management Limited/DEEP Conflict Climate Bangladesh/")

# Import SPEI data for 2010-2020 covering all of Bangladesh on a regular 0.5 degrees grid
spei_geo <- read.csv("Code/SPEI_2010_20_BGD.csv")

# Load shapefile with administrative level 4 for Bangladesh
bgd_admin4 <- read_sf("Data/Administrative boundary shapefiles/bgd_adm_bbs_20201113_shp/bgd_adm_bbs_20201113_SHP/bgd_admbnda_adm4_bbs_20201113.shp")

# Load BIHS location data
BIHS_admin <- read_dta("Data/BIHS data/BIHSlocations.dta")

# Remove unnecessary columns
spei_geo <- spei_geo[,2:14]

# Turn data frame with SPEI values into a raster - this creates a "brick" where values of SPEI for all years are on top of one another
raster_spei <- rasterFromXYZ(spei_geo)

# Make sure the CRS is consistent with shapefile
crs(raster_spei) <- "+proj=longlat +datum=WGS84 +no_defs"

```


## Create a shapefile with corrected union codes that match BIHS ##

For all 14 cases, look at the Stata dta file with BIHS locations that also has labels with names in English. On that basis, find the corresponding ADM4_PCODE in the shapefile, which is different.

1. BIHS code: 193617, BIHS name: Bitikandi, Shapefile ADM4_PCODE: 199417
2. BIHS code: 198104, BIHS name: Akubpur, Shapefile ADM4_PCODE: 198110
3. BIHS code: 338609, BIHS name: Barmi, Shapefile ADM4_PCODE: 338621
4. BIHS code: 354306, BIHS name: Bethuri, Shapefile ADM4_PCODE: 354311
5. BIHS code: 827395, BIHS name: Saorail, Shapefile ADM4_PCODE: 824795
6. BIHS code: 935780, BIHS name: Musuddi, Shapefile ADM4_PCODE: 932580
7. BIHS code: 102056, BIHS name: Majhira , Shapefile ADM4_PCODE: 108556
8. BIHS code: 105409, BIHS name: Bir Kedar, Shapefile ADM4_PCODE: 105413
9. BIHS code: 386109, BIHS name: Alampur, Shapefile ADM4_PCODE: 386115
10. BIHS code: 696312, BIHS name: Bara Harishpur, Shapefile ADM4_PCODE: 696320
11. BIHS code: 496106, BIHS name: Ballabher Khas, Shapefile ADM4_PCODE: 496111
12. BIHS code: 855816, BIHS name: Bara Hazratpur, Shapefile ADM4_PCODE: 855826
13. BIHS code: 857309, BIHS name: Annadanagar, Shapefile ADM4_PCODE: 857317
14. BIHS code: 857650, BIHS name:  Madankhali, Shapefile ADM4_PCODE: 857656


```{r BIHS locations SPEI }
# Rename the code identifiers for administrative levels - remove BD
bgd_admin4$ADM4_PCODE <- substr(bgd_admin4$ADM4_PCODE, 3, 10)

# Remove the first 2 digits from the union code (those refer to admin level 1)
bgd_admin4$ADM4_PCODE <- substr(bgd_admin4$ADM4_PCODE, 3, 10)

# This is correct for most cases. However, whenever "0" is the first digit, we should delete it 

# Change to numeric - this will drop 0 at the front 
bgd_admin4$ADM4_PCODE <- as.numeric(bgd_admin4$ADM4_PCODE)

# Back to character 
bgd_admin4$ADM4_PCODE <- as.character(bgd_admin4$ADM4_PCODE)

# Rename columns in BIHS_admin before the join 
names(BIHS_admin)[names(BIHS_admin) == "uncode"] <- "ADM4_PCODE"

# Back to character for compatibility of the merge
BIHS_admin$ADM4_PCODE <- as.character(BIHS_admin$ADM4_PCODE)

# Create another data frame with values from the Bangladesh shapefile to replace with new matching codes. Here I replace shapefile codes with codes from BIHS.
bgd_admin4 <- bgd_admin4 %>% 
  mutate(uncode = case_when(
    ADM4_PCODE == "199417" ~ "193617",      # 1
    ADM4_PCODE == "198110" ~ "198104",      # 2
    ADM4_PCODE == "338621" ~ "338609",      # 3
    ADM4_PCODE == "354311" ~ "354306",      # 4
    ADM4_PCODE == "824795" ~ "827395",      # 5
    ADM4_PCODE == "932580" ~ "935780",      # 6
    ADM4_PCODE == "108556" ~ "102056",      # 7
    ADM4_PCODE == "105413" ~ "105409",      # 8
    ADM4_PCODE == "386115" ~ "386109",      # 9
    ADM4_PCODE == "696320" ~ "696312",      # 10
    ADM4_PCODE == "496111" ~ "496106",      # 11
    ADM4_PCODE == "855826" ~ "855816",      # 12
    ADM4_PCODE == "857317" ~ "857309",      # 13
    ADM4_PCODE == "857656" ~ "857650"))     # 14

# Transfer all other non-missing codes 
bgd_admin4$uncode[is.na(bgd_admin4$uncode)] <- bgd_admin4$ADM4_PCODE

# Drop the original column and rename the new one 
bgd_admin4<- bgd_admin4[ , !names(bgd_admin4) %in% 
    c("ADM4_PCODE")]
names(bgd_admin4)[names(bgd_admin4) == 'uncode'] <- "ADM4_PCODE"

# Move admin 4 code to the front 
bgd_admin4 <- bgd_admin4 %>% 
  relocate("ADM4_PCODE")

# Save shapefile - this will create a shapefile consistent with BIHS codes for future reference 
st_write(bgd_admin4, "C:/Users/idabr/OneDrive - Oxford Policy Management Limited/DEEP Conflict Climate Bangladesh/Data/Harmonised shapefile admin 4bgd_admin4_bihs_harmonised.shp")
```

# Calculate the mean value of SPEI for each union 4 #

```{r}
# Calculate the mean value of SPEI in each union (administrative level 4) - keep identifiers for admin level 4
bgd_spei <- exact_extract(raster_spei, bgd_admin4, fun = c('mean'), append_cols=c("ADM4_PCODE", "ADM4_EN"))

# Check that there are no missing values -3 NA values (coastal areas which the grid did not cover)
summary(bgd_spei)
```

Match with BIHS data based on corrected codes

```{r union codes}
# Left join - all codes should match 
BIHS_drought <- left_join(BIHS_admin, bgd_spei_newcodes)

# Remove "mean" from column names
names(BIHS_drought) <- sub("^mean.", "", names(BIHS_drought))

# Export BIHS data matched with flood data a csv file where the "ADM4_PCODE" column corresponds to the "uncode" column from the BIHS data. 
write.csv(BIHS_drought, "BIHS_drought.csv")

# Check for missing values - no NAs :)
summary(BIHS_drought)

```
